% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AF20150710.R
\name{AF.cc}
\alias{AF.cc}
\title{Attributable fraction mached or non-matched case-control sampling design.}
\usage{
AF.cc(formula, data, exposure, clusterid, sampling.design = "non.matched")
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model used for adjusting for confounders. The independent variables should be specified as the exposure and confounders. The dependent variable should be specified as the outcome of interest. The formula is used to fit a logistic regression by glm for non-matched cc and conditional logistic regression by gee (in package drgee) for macthed cc. For details see documentation on glm or gee.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which the function is called.}

\item{exposure}{the exposure variable. Must be binary (0/1).}

\item{sampling.design}{a string which specify if the sampling is matched or non-matched case-control. Default setting is non-matched.}
}
\value{
\item{AF.est}{estimated attributable fraction}

\item{AF.var}{estimated variance of the AF estimate (\code{AF.est})}

\item{log.or}{estimated log odds ratio}
}
\description{
\code{AF.cc} estimate the attributable fraction for data from a mached or non-matched case-control sampling design.
}
\details{
\code{Af.cc} estimate the attributable fraction for binary outcomes
under the scenario of an elimination of the exposure adjusted for confounders
by logistic regression for unmatched case-control and conditional logistic regression for matched case-control.
The estimation is an approximation based on the assumption that the outcome is rare.
Let the AF be defined as
\deqn{AF = 1 - Pr(Y_0 = 1) / Pr(Y = 1)}
where \eqn{Pr(Y_0 = 1)} denote the counterfactual probability of the outcome if
the exposure would be eliminated from the population. Under the assumption that Z
is the only confounder the probability \eqn{Pr(Y_0 = 1)} can be expressed as
\deqn{E_z{Pr(Y = 1 | X = 0, Z)}}
Using Baye's theorem this imply that the AF can be defined as
  \deqn{AF = 1-\frac{E_z\{Pr(Y=1\mid X=0,Z)\}}{Pr(Y=1)}=1-E{RR^{-X}(Z)|Y=1}}{%
AF = 1 - E_z{Pr( Y=1 | X=0, Z)}/ Pr(Y = 1) = 1 - E{RR^{-X}(Z) | Y = 1}}
where \eqn{E{RR^{-X}(Z) | Y = 1}} can be estimated by the relative risk.
Moreover, the relative risk can be approximated by the odds ratio if the outcome is rare.
The odds ratio is estimated by logistic regression or conditional logistic regression. Thus,
\deqn{1 - E{RR^{-X}(Z) | Y = 1} \approx 1 - E{OR^{-X}(Z) | Y = 1}}
where \eqn{E{OR^{-X}(Z) | Y = 1}} is the sum of the parameters depending on the exposure X.
}
\author{
Elisabeth Dahlqwist, Arvid Sjolander
}
\references{
Bruzzi, P., Green, S. B., Byar, D., Brinton, L. A., and Schairer, C. (1985). Estimating the population attributable risk for multiple risk factors using case-control data. \emph{American Journal of Epidemiology} \bold{122}, 904-914.
}
\seealso{
\code{\link{glm}} and \code{\link{gee}} used for estimating the conditional logistic regression for mached case-control.
}

