% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AF20150710.R
\name{AF.ch}
\alias{AF.ch}
\title{Attributable fraction function from cohort sampling design with time-to-event outcomes.}
\usage{
AF.ch(formula, data, exposure, ties = "breslow", time.sequence, clusterid)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function. A symbolic description of the model used for adjusting for confounders. The independent variables should be specified as the exposure and confounders. The dependent variable should be specified as the outcome of interest. The formula is used to fit a Cox Proportional Hazards model in the survival package. For details see documentation on coxph.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which the function is called.}

\item{exposure}{the exposure variable. Must be binary (0/1).}

\item{ties}{a character string specifying the method for tie handling. If there are no tied death times all the methods are equivalent. Use the Breslow method by default.}

\item{time.sequence}{a single value or vector of time points of which the attrubutable fraction function will be calculated. If not specified the death times will be used as default.}

\item{clusterid}{for clustered data specify the variable in the data frame which is the cluster id in order to calculate cluster robust standard errors.}
}
\value{
\item{AF.est}{estimated attributable fraction function for every time point specified by \code{time.sequence}}

\item{AF.var}{estimated variance of the AF estimate (\code{AF.est}) for every time point specified by \code{time.sequence}}

\item{St.est}{estimated factual survival function}

\item{St.var}{estimated variance of the estimate \code{St.est}}

\item{St0.est}{estimated counterfactual survival function if exposure would be eliminated}

\item{St0.var}{estimated variance of the estimate \code{St.est}}
}
\description{
\code{AF.ch} estimate the model-based adjusted attributable fraction function for data from a cohort sampling design with time-to-event outcomes.
}
\details{
\code{Af.ch} estimate the attributable fraction for time-to-event outcomes
under the scenario of an elimination of the exposure adjusted for confounders
by Cox Proportional Hazards model. Let the Attributable fraction function be defined as
\deqn{AF = 1 - ( 1 - St0(t) ) / ( 1 - St(t) )}
where \eqn{St0(t)} denote the counterfactual survival function for the event if
the exposure would be eliminated from the population at baseline. \eqn{St(t)} is the factual survival function.
\eqn{St0(t)} is estimated by adjusting for confounders at baseline using the Cox Proportional Hazards model.
\eqn{St0(t)} is denoted as \code{St0.est} in the output and \eqn{St(t)}
is denoted as \code{St.est}.
}
\author{
Elisabeth Dahlqwist, Arvid Sjolander
}
\references{
Chen, L., Lin, D. Y., and Zeng, D. (2010). Attributable fraction functions for censored event times. \emph{Biometrika} \bold{97}, 713-726.

Sjolander, A. and Vansteelandt, S. (2014). Doubly robust estimation of attributable fractions in survival analysis. \emph{Statistical Methods in Medical Research}.
}
\seealso{
\code{\link{coxph}} and \code{\link{Surv}}.
}

