% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AF20150710.R
\name{AF.cs}
\alias{AF.cs}
\title{Attributable fraction cross-sectional sampling design}
\usage{
AF.cs(formula, data, exposure, clusterid)
}
\arguments{
\item{formula}{an object of class "\code{\link{formula}}" (or one that can be coerced to that class): a symbolic description of the model used for adjusting for confounders. The independent variables should be specified as the exposure and confounders. The dependent variable should be specified as the outcome of interest. The formula is used to fit a logistic regression by \code{\link{glm}}.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which the function is called.}

\item{exposure}{the exposure variable. Must be binary (0/1).}

\item{clusterid}{for clustered data specify the variable in the data frame which is the cluster id in order to calculate cluster robust standard errors.}
}
\value{
\item{AF.est}{estimated attributable fraction}

\item{AF.var}{estimated variance of the AF estimate (\code{AF.est})}

\item{P}{estimated factual proportion of cases}

\item{P.var}{estimated variance of the estimate \code{P}}

\item{P0}{estimated counterfactual proportion of cases if exposure would be eliminated}

\item{P0.var}{estimated variance of the estimate \code{P0}}
}
\description{
\code{AF.cs} estimate the model-based adjusted attributable fraction for data from a cross-sectional sampling design.
}
\details{
\code{Af.cs} estimate the attributable fraction for binary outcomes
under the scenario of an elimination of the exposure adjusted for confounders
by logistic regression. Let the AF be defined as
\deqn{AF = 1 - Pr(Y_0 = 1) / Pr(Y = 1)}
where \eqn{Pr(Y_0 = 1)} denote the counterfactual probability of the outcome if
the exposure would be eliminated from the population.
By adjusting the attributable fraction for confounders the counterfactual probablity
 \eqn{Pr(Y_0 = 1)} can be estimated. \eqn{Pr(Y_0 = 1)} is denoted as \code{P0} in the
 output and \eqn{Pr(Y = 1)} is denoted as \code{P}.
}
\author{
Elisabeth Dahlqwist, Arvid Sjolander
}
\references{
Bruzzi, P., Green, S. B., Byar, D., Brinton, L. A., and Schairer, C. (1985). Estimating the population attributable risk for multiple risk factors using case-control data. \emph{American Journal of Epidemiology} \bold{122}, 904-914.

Greenland, S. (1984). Bias in methods for deriving standardized morbidity ratio and attributable fraction estimates. \emph{Statistics in medicine} \bold{3}, 131-141.

Greenland, S. and Drescher, K. (1993). Maximum Likelihood Estimation of the Attributable Fraction from logistic Models. \emph{Biometrics} \bold{49}, 865-872.
}
\seealso{
\code{\link{glm}}.
}

