% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AF20150710.R
\name{AF.cc}
\alias{AF.cc}
\title{Attributable fraction mached or non-matched case-control sampling design.}
\usage{
AF.cc(formula, data, exposure, clusterid, sampling.design = "non.matched")
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model used for adjusting for confounders. The independent variables should be specified as the exposure and confounders. The dependent variable should be specified as the outcome of interest. The formula is used to fit a logistic regression by glm for non-matched cc and conditional logistic regression by gee (in package drgee) for macthed cc. For details see documentation on glm or gee.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which the function is called.}

\item{exposure}{the exposure variable. Must be binary (0/1).}

\item{sampling.design}{a string which specify if the sampling is matched or non-matched case-control. Default setting is non-matched.}
}
\value{
An estimated attributable fraction and standard errors for the estimate.
}
\description{
\code{AF.cc} estimate the attributable fraction for data from a mached or non-matched case-control sampling design.
}
\details{
Note that the estimation is an approximation based on the assumption that the outcome is rare.
}
\author{
Elisabeth Dahlqwist, Arvid Sjölander
}
\seealso{
\code{\link{glm}} and \code{\link{gee}} used for estimating the conditional logistic regression for mached case-control.
}

