% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AF20150710.R
\name{AF.ch}
\alias{AF.ch}
\title{Attributable fraction function from cohort sampling design with time-to-event outcomes.}
\usage{
AF.ch(formula, data, exposure, ties = "breslow", time.sequence, clusterid)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function. A symbolic description of the model used for adjusting for confounders. The independent variables should be specified as the exposure and confounders. The dependent variable should be specified as the outcome of interest. The formula is used to fit a Cox Proportional Hazards model in the survival package. For details see documentation on coxph.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which the function is called.}

\item{exposure}{the exposure variable. Must be binary (0/1).}

\item{ties}{a character string specifying the method for tie handling. If there are no tied death times all the methods are equivalent. Use the Breslow method by default.}

\item{time.sequence}{a single value or vector of time points of which the attrubutable fraction function will be calculated. If not specified the death times will be used as default.}

\item{clusterid}{for clustered data specify the variable in the data frame which is the cluster id in order to calculate cluster robust standard errors.}
}
\value{
An estimated attributable fraction function and standard errors for the estimate for every time point as specified by time.sequence.
}
\description{
Estimate the model-based adjusted attributable fraction function for data from a cohort sampling design with time-to-event outcomes.
}
\details{
Note that the estimation is an approximation based on the assumption that the outcome is rare.
}
\author{
Elisabeth Dahlqwist, Arvid Sjölander
}
\seealso{
glm and gee in package "drgee"
}

